refactor(core): rename get-entry-from-cli-or-config to get-entries-from-cli-or-config.ts
