+++ packages/core/src/lib/cli/export-data.ts
@@ -1,4 +1,4 @@
-import { getEntryFromCliOrConfig } from './internal/get-entry-from-cli-or-config';
+import { getEntryFromCliOrConfig } from './internal/get-entries-from-cli-or-config';
 import { cli } from './cli';
 import { getProjectData } from '../api/get-project-data';
 import getFs from '../fs/getFs';
@@ -7,6 +7,8 @@ export function exportData(...args: string[]): void {
   const fs = getFs();
   const entryFile = getEntryFromCliOrConfig(args[0], false);
 
-  const data = getProjectData(entryFile, fs.cwd(), { includeExternalLibraries: true });
+  const data = getProjectData(entryFile, fs.cwd(), {
+    includeExternalLibraries: true,
+  });
   cli.log(JSON.stringify(data, null, '  '));
 }
